Introduce CI mode
Two new configs are introduced:
1. enable_ci: bool
2. package_lock_file: string

When enable_ci is true, cmd updater will seek the package lock file
to determine the package version to update. As long as it is different
from the one in the local repository, the update will proceed.

The package_lock_file config is ignored when enable_ci is false