Pass full command name to executables
Use the ${$PREFIX}_COMMAND_NAME environment variable to pass the full
command to the executable files when they are run.

The full command would be the app name, followed optionally by the group
name, and finally the command name that was configured in the manifest
to execute the command.

This can be used, for instance, by programs to output a meaningful usage
message that refers to how the program was actually executed.